// at-backend/src/common/enums/group-code.enum.ts
export enum GroupCode {
  GROUP_I = 'GROUP_I',
  GROUP_II = 'GROUP_II',
  GROUP_II_BIFOS = 'GROUP_II_BIFOS',
  GROUP_II_SUCRA = 'GROUP_II_SUCRA',
  GROUP_III = 'GROUP_III',
  GROUP_III_LAX = 'GROUP_III_LAX',
  GROUP_III_SAL = 'GROUP_III_SAL',
  GROUP_III_ESTAT = 'GROUP_III_ESTAT',
  GROUP_III_DIU = 'GROUP_III_DIU',
  GROUP_III_MET = 'GROUP_III_MET',
  GROUP_III_SUL = 'GROUP_III_SUL',
  GROUP_III_SUL2 = 'GROUP_III_SUL2',
  GROUP_III_PROC = 'GROUP_III_PROC',
  GROUP_I_SED = 'GROUP_I_SED',
  GROUP_III_CALC = 'GROUP_III_CALC',
  GROUP_III_FER = 'GROUP_III_FER',
  GROUP_DELTA = 'GROUP_DELTA',
  GROUP_INSUL_ULTRA = 'GROUP_INSUL_ULTRA',
  GROUP_INSUL_RAPIDA = 'GROUP_INSUL_RAPIDA',
  GROUP_INSUL_INTER = 'GROUP_INSUL_INTER',
  GROUP_INSUL_LONGA = 'GROUP_INSUL_LONGA'
}